Move enums to separate folder
